# ggml-zdnn: rm user mapped buffers